fixes + useMemo cases
