+++ src/components/genre-list/genre-list.tsx
@@ -1,5 +1,5 @@
 import {useNavigate, useSearchParams} from 'react-router-dom';
-import {useEffect} from 'react';
+import {useEffect, useMemo} from 'react';
 import {useAppDispatch, useAppSelector} from '../../hooks/redux-typed-hooks';
 import {DefaultFilmGenre} from '../../const.ts';
 import {changeGenre} from '../../store/main/main-slice.ts';
@@ -11,7 +11,7 @@ export const GenreList = () => {
   const dispatch = useAppDispatch();
   const films = useAppSelector((state) => state.main.films);
 
-  const uniqueGenres = [DefaultFilmGenre, ...new Set(films.map((x) => x.genre))];
+  const uniqueGenres = useMemo(() => [DefaultFilmGenre, ...new Set(films.map((x) => x.genre))], [films]);
   const [searchParams] = useSearchParams();
 
   useEffect(() => {
